User table update for rating decimal and user stored procedures

+++ PhoneDb/dbo/Tables/Users.sql
@@ -5,7 +5,7 @@
 	[UserName] NVARCHAR(150) NOT NULL,
 	[Password] NVARCHAR(256) NOT NULL, 
 	[Description] NVARCHAR(256) NULL,
-	[Rating] INT NULL,
+	[Rating] DECIMAL(3, 2) NULL,
 	[PhoneNumber] NVARCHAR(100) NULL,
 	[Phones_sold] int NOT NULL DEFAULT 0,
 	[EmailConfirmed] bit NOT NULL DEFAULT 0,
